Fix and clean
