--- src/executor/operator/physical_match_homebrewed.cppm
@@ -1,34 +1,0 @@
-// Copyright(C) 2023 InfiniFlow, Inc. All rights reserved.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     https://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-module;
-
-export module physical_match_homebrewed;
-
-import stl;
-import query_context;
-import operator_state;
-import match_expression;
-import base_table_ref;
-import data_type;
-
-namespace infinity {
-
-export bool ExecuteInnerHomebrewed(QueryContext *query_context,
-                                   OperatorState *operator_state,
-                                   SharedPtr<BaseTableRef> &base_table_ref_,
-                                   SharedPtr<MatchExpression> &match_expr_,
-                                   Vector<SharedPtr<DataType>> OutputTypes);
-
-} // namespace infinity
